# server/openapi.yaml
openapi: 3.1.0
info:
  title: Finlo API
  version: 0.1.0
  description: Shared REST contract for Finlo web and mobile clients.
servers:
  - url: http://localhost:8080
paths:
...
  /api/goals:
    get:
      operationId: listGoals
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Goals
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Goal' } }
    post:
      operationId: createGoal
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGoalInput' }
      responses:
        '201':
          description: Goal created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Goal' }
  /api/goals/{id}:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    patch:
      operationId: updateGoal
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateGoalInput' }
      responses:
        '200':
          description: Goal updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Goal' }
    delete:
      operationId: deleteGoal
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Goal deleted }
...
  /api/integrations/monobank:
    get:
      operationId: getMonobankConnection
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current Monobank connection or null
          content:
            application/json:
              schema: { oneOf: [{ $ref: '#/components/schemas/MonobankConnection' }, { type: 'null' }] }
  /api/integrations/monobank/preview:
    post:
      operationId: previewMonobankConnection
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MonobankPreviewInput' }
      responses:
        '200':
          description: Validated Monobank resources
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonobankPreview' }
        default: { $ref: '#/components/responses/Error' }
  /api/integrations/monobank/complete:
    post:
      operationId: completeMonobankConnection
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteMonobankInput' }
      responses:
        '200':
          description: Monobank connection state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonobankConnection' }
        default: { $ref: '#/components/responses/Error' }
  /api/integrations/monobank/webhook:
    put:
      operationId: registerMonobankWebhook
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current Monobank connection after webhook registration
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonobankConnection' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      operationId: disconnectMonobankWebhook
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Monobank connection after webhook removal; imported data remains intact
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonobankConnection' }
        default: { $ref: '#/components/responses/Error' }
  /webhooks/monobank/{secret}:
    parameters:
      - in: path
        name: secret
        required: true
        schema: { type: string, minLength: 64, maxLength: 64 }
    get:
      operationId: validateMonobankWebhook
      responses:
        '200': { description: Webhook URL accepted }
    post:
      operationId: receiveMonobankWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Event processed }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: finlo_refresh }
  parameters:
    ID: { in: path, name: id, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/APIError' }
  schemas:
    SupportedCurrency:
      type: string
      enum: [UAH, USD, EUR]
...
    APIError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    Goal:
      type: object
      required: [id, name, current_minor, currency, source, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        current_minor: { type: integer, format: int64, minimum: 0 }
        target_minor: { type: integer, format: int64, minimum: 1 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        source: { type: string, enum: [manual, monobank] }
        external_id: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateGoalInput:
      type: object
      required: [name, current_minor, currency]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        current_minor: { type: integer, format: int64, minimum: 0 }
        target_minor: { type: integer, format: int64, minimum: 1 }
        currency: { $ref: '#/components/schemas/SupportedCurrency' }
    UpdateGoalInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        current_minor: { type: integer, format: int64, minimum: 0 }
        target_minor: { type: integer, format: int64, minimum: 1 }
        clear_target: { type: boolean }
...
    MonobankPreviewInput:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 1, maxLength: 500 }
    MonobankPreviewAccount:
      type: object
      required: [id, name, type, currency, balance_minor, credit_limit_minor, masked_pan, iban]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        currency: { type: string }
        balance_minor: { type: integer, format: int64 }
        credit_limit_minor: { type: integer, format: int64 }
        masked_pan: { type: array, items: { type: string } }
        iban: { type: string }
    MonobankPreviewJar:
      type: object
      required: [id, title, description, currency, balance_minor]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        currency: { type: string }
        balance_minor: { type: integer, format: int64 }
        target_minor: { type: integer, format: int64, minimum: 1 }
    MonobankPreview:
      type: object
      required: [connection_id, client_name, accounts, jars]
      properties:
        connection_id: { type: string, format: uuid }
        client_name: { type: string }
        accounts: { type: array, items: { $ref: '#/components/schemas/MonobankPreviewAccount' } }
        jars: { type: array, items: { $ref: '#/components/schemas/MonobankPreviewJar' } }
    CompleteMonobankInput:
      type: object
      required: [connection_id, account_ids, jar_ids]
      properties:
        connection_id: { type: string, format: uuid }
        account_ids: { type: array, items: { type: string } }
        jar_ids: { type: array, items: { type: string } }
    MonobankConnection:
      type: object
      required: [id, status, client_name, webhook_configured, account_count, jar_count]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [pending, active, webhook_error, disconnected] }
        client_name: { type: string }
        webhook_configured: { type: boolean }
        account_count: { type: integer, format: int64 }
        jar_count: { type: integer, format: int64 }
        last_error: { type: string }
        connected_at: { type: string, format: date-time }
...